requirements.txt
streamlit
st-annotated-text
numpy
pandas
google.generativeai
google-genai #for keyword extract
matplotlib
psycopg2-binary
sqlalchemy
beautifulsoup4
gTTS
rake-nltk
pydantic